Cheats bypass

+++ TeamFortress2/TeamFortress2/Features/Visuals/Visuals.cpp
@@ -65,7 +65,7 @@ void CVisuals::AddToEventLog(const char* string...) {
 
 void CVisuals::RunEventLogs()
 {
-	for (auto i = 0; i < vecEventVector.size(); i++) {
+	for (auto i = 0; i < (int)vecEventVector.size(); i++) {
 		auto log = vecEventVector[i];
 		auto time_delta = fabs(g_Interfaces.GlobalVars->realtime - log.flTime);
 
@@ -273,3 +273,4 @@ void CVisuals::RestoreWorldModulation()
 	ApplyModulation({ 255, 255, 255, 255 });
 	bWorldIsModulated = false;
 }
+
